fix(enums): Single cases overwriting the global enum type
We now check more strictly that the transformer only tries to transform actual enums and not potential enums.

Single enum cases also inherit from the ObjectType class and we cannot make them have the enum class name.
Otherwise, they will overwrite the enum class type definition with their single case, omitting the other potential cases.

We now only set the name, when the full enum is actually used as a type.

The transformer was also adjusted to not return unnecessary union types when it only encapsulates a single type.